Modify subproj Makefiles to Include RPI-Specific Build Procedures
Each Makefile was given an extra 'rpi' command to build only for
the raspberry pi. In addition, a 'clean' command was implemented.
Both of these changes were made with amiibrOS_post-build config in
mind: it can now more easily build and copy only the needed rpi
files.

Also updated versions of the submodules.